+++ ds-license/src/test/java/dk/kb/license/facade/RightsModuleFacadeTest.java
@@ -5,7 +5,7 @@ import dk.kb.license.model.v1.*;
 import dk.kb.license.solr.SolrServerClient;
 import dk.kb.license.storage.BaseModuleStorage;
 import dk.kb.license.storage.RightsModuleStorageForUnitTest;
-import dk.kb.license.storage.UnitTestUtil;
+import dk.kb.license.storage.DsStorageUnitTestUtil;
 import dk.kb.license.util.H2DbUtil;
 import dk.kb.license.webservice.KBAuthorizationInterceptor;
 import dk.kb.util.webservice.exception.InternalServiceException;
@@ -38,7 +38,7 @@ import java.util.Locale;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
-public class RightsModuleFacadeTest extends UnitTestUtil {
+public class RightsModuleFacadeTest extends DsStorageUnitTestUtil {
     protected static RightsModuleStorageForUnitTest storage = null;
     static MockedStatic<JAXRSUtils> mocked;