+++ ds-license/src/test/java/dk/kb/license/storage/DsStorageUnitTestUtil.java
@@ -11,7 +11,7 @@ import org.slf4j.LoggerFactory;
  * 1) Create a h2 database for unittests with schema defined
  * 2) Load the Yaml property files.
  */
-public abstract class UnitTestUtil {
+public abstract class DsStorageUnitTestUtil {
     protected static final String DRIVER = "org.h2.Driver";
 
     //We need the relative location. This works both in IDE's and Maven.
@@ -20,7 +20,7 @@ public abstract class UnitTestUtil {
     protected static final String USERNAME = "";
     protected static final String PASSWORD = "";
 
-    private static final Logger log = LoggerFactory.getLogger(UnitTestUtil.class);
+    private static final Logger log = LoggerFactory.getLogger(DsStorageUnitTestUtil.class);
 
     @AfterAll
     public static void afterClass() {