+++ ds-storage/src/test/java/dk/kb/storage/facade/DsStorageFacadeTestDsStorage.java
@@ -6,7 +6,7 @@ import dk.kb.storage.model.v1.RecordTypeDto;
 import dk.kb.storage.storage.BaseModuleStorage;
 import dk.kb.storage.storage.DsStorage;
 import dk.kb.storage.storage.DsStorageForUnitTest;
-import dk.kb.storage.storage.UnitTestUtil;
+import dk.kb.storage.util.DsStorageUnitTestUtil;
 import dk.kb.storage.util.H2DbUtil;
 import dk.kb.util.webservice.exception.InternalServiceException;
 import org.junit.jupiter.api.AfterAll;
@@ -22,7 +22,7 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
-public class DsStorageFacadeTest extends UnitTestUtil {
+public class DsStorageFacadeTestDsStorage extends DsStorageUnitTestUtil {
     protected static DsStorageForUnitTest storage = null;
 
     @BeforeAll